+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - unreleased
+### Fixed
+- MX records now get a default prio=0 if not set
 
 ## [1.3.2] - 2016-11-20
 ### Added